split clipboard helper to a few ones
